refactor(bayn): restore one-way cycle layering
