+++ services/bayn/src/cycle-runner/recovery-decision-binding.ts
@@ -1,14 +1,17 @@
 import { Result } from 'effect'
 
-import { CycleState, CycleTerminalReason, type AutonomousCycle, type CycleCompletionState } from '../cycle'
+import type { AutonomousCycle } from '../cycle'
 import type { CycleDecisionDocument } from '../shadow-decision-contract'
-import { TargetPlanReason, TargetPlanStatus, type BlockedTargetPlanReason } from '../target-planner'
+import { TargetPlanStatus } from '../target-planner'
 import {
   selectRecoveryFailure,
   validateDecisionFailure,
   type CycleRecoveryFailure,
   type CycleRecoverySelection,
 } from './recovery-model'
+import { cycleCompletionStateForTargetPlan, cycleTerminalReasonForBlockedTargetPlan } from './recovery-decisions'
+
+export { cycleCompletionStateForTargetPlan, cycleTerminalReasonForBlockedTargetPlan } from './recovery-decisions'
 
 const validateDecisionBinding = (
   cycle: AutonomousCycle,
@@ -56,42 +59,6 @@ const validateDecisionBinding = (
         ),
       )
     : Result.succeed(undefined)
-}
-
-export const cycleTerminalReasonForBlockedTargetPlan = (reason: BlockedTargetPlanReason): CycleTerminalReason => {
-  switch (reason) {
-    case TargetPlanReason.SubmissionCutoffReached:
-      return CycleTerminalReason.MissedSubmission
-    case TargetPlanReason.IdentityMismatch:
-      return CycleTerminalReason.ProvenanceMismatch
-    case TargetPlanReason.InputMismatch:
-      return CycleTerminalReason.DataInvalid
-    case TargetPlanReason.InputStale:
-      return CycleTerminalReason.DataStale
-    case TargetPlanReason.ReconciliationNotExact:
-      return CycleTerminalReason.Reconciliation
-    case TargetPlanReason.AccountNotActive:
-      return CycleTerminalReason.BrokerDisabled
-    case TargetPlanReason.UnknownOrder:
-    case TargetPlanReason.UnresolvedOrder:
-      return CycleTerminalReason.UnresolvedMutation
-    case TargetPlanReason.BelowMinimumBuyNotional:
-    case TargetPlanReason.InsufficientBuyingPower:
-    case TargetPlanReason.NonPositiveEquity:
-    case TargetPlanReason.ShortPositionNotAllowed:
-      return CycleTerminalReason.Risk
-  }
-}
-
-export const cycleCompletionStateForTargetPlan = (
-  status: TargetPlanStatus.Planned | TargetPlanStatus.NoTrade,
-): CycleCompletionState => {
-  switch (status) {
-    case TargetPlanStatus.Planned:
-      return CycleState.Completed
-    case TargetPlanStatus.NoTrade:
-      return CycleState.NoTrade
-  }
 }
 
 export const selectBoundDecision = (
